renamed env to constants
