+++ manage.py
@@ -11,8 +11,8 @@ def main():
     except ImportError as exc:
         raise ImportError(
             "Couldn't import Django. Are you sure it's installed and "
-            "available on your PYTHONPATH environment variable? Did you "
-            "forget to activate a virtual environment?"
+            "available on your PYTHONPATH constants variable? Did you "
+            "forget to activate a virtual constants?"
         ) from exc
     execute_from_command_line(sys.argv)
