--- master_service/utilities/send_error.py
@@ -1,7 +1,0 @@
-import requests
-
-
-# utility file for sending structured and unified error messages to frontend
-
-def send_error(http_status, msg):
-    return 'ok'
